requirements.txt
pyTelegramBotAPI==4.12.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3 